+++ ejo-core.php
@@ -58,6 +58,9 @@ final class EJO_Core
         //* Define Includes Directory
         if ( ! defined( 'THEME_INC_DIR' ) ) define( 'THEME_INC_DIR', THEME_DIR . 'includes/' );
         if ( ! defined( 'THEME_INC_URI' ) ) define( 'THEME_INC_URI', THEME_URI . 'includes/' );
+        
+        // Composer Vendor Dir
+        if ( ! defined( 'THEME_VENDOR_DIR' ) ) define( 'THEME_VENDOR_DIR', THEME_DIR . 'vendor/' );
 
         //* Set Version
         if ( ! defined( 'THEME_VERSION' ) ) define( 'THEME_VERSION', wp_get_theme()->get( 'Version' ) );
@@ -70,7 +73,7 @@ final class EJO_Core
         if ( ! defined( 'THEME_VENDOR_URI' ) ) define( 'THEME_VENDOR_URI', THEME_URI . 'assets/dist/vendor/' );    
 
         // TODO: Better to get the path of ejo-core automatically in stead of manually...
-        $relative_framework_path = trailingslashit( apply_filters( 'ejocore_relative_framework_path', 'includes/vendor/' ) );
+        $relative_framework_path = trailingslashit( apply_filters( 'ejocore_relative_framework_path', 'vendor/ejoweb/' ) );
 
         // Sets the path to the core framework directory.
         if ( ! defined( 'EJO_CORE_DIR' ) )