+++ CHANGES.md
@@ -1,3 +1,6 @@
+- Added vendor directory global
+- Added composer configuration
+
 0.2.1
 - Included royalslider to assets